+++ CHANGELOG.md
@@ -9,6 +9,9 @@ ENHANCEMENTS:
 * mdb: add `conn_limit` attribute to `user` entity in `yandex_mdb_postgresql_cluster` resource and data source
 * mdb: add `config` section in `yandex_mdb_clickhouse_cluster` resource and data source
 
+FEATURES:
+* storage: `yandex_storage_bucket` resource supports lifecycle configuration
+
 ## 0.47.0 (November 10, 2020)
 BUG FIXES:
 * kms: fix import operation